smart-waste-system-with-models/requirements.txt
# streamlit==1.33.0
# pandas==2.2.2
# numpy==1.26.4
# pillow==10.3.0
# opencv-python==4.9.0.80

# Core dependencies
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80

# TensorFlow untuk Apple Silicon (M4)
tensorflow==2.16.1  # Versi terbaru dengan native ARM64 support
tensorflow-metal==1.1.0  # Untuk GPU acceleration via Metal

# Keras (bundled with TensorFlow 2.16+)

# Data Visualization
matplotlib==3.8.4
plotly==5.22.0
seaborn==0.13.2

# Image Processing
scikit-image==0.22.0
imageio==2.34.0

# Utilities
scikit-learn==1.5.0
tqdm==4.66.4
pyyaml==6.0.1

# MQTT for IoT
paho-mqtt==2.1.0

# For TensorFlow
protobuf==3.20.3
h5py==3.11.0